enable loading model weights from internet
This commit adds a Weights object which stores information about a
trained model. This includes the url to weights, number of output
classes, patch size, patch spacing, and metadata.

In the future, we might want to add the human-readable classes list to
the Weights object, because it makes sense for them to be kept with the
weights.