ADD: TILs inceptionv4 + docker images for apps (#21)
* add tcga-luad-v1 resnet34 weights

* resnet34: check that missing keys are missing

* add tcga-prad-v1 resnet34 weights + fix file name for tcga-luad-v1 resnet34 weights

* add script to convert vgg16 TIL model

* add inceptionv4 conversion script + add inceptionv4 without bn

The TF Slim model was trained without batchnorm. So this commit adds an
inceptionv4 implementation that does not use batchnorm. Bias is also
used in the inceptionv4 without batchnorm (the typical inceptionV4 uses
bathnorm but not bias).

* add inceptionv4 weights for tils

* add TCGA-TILs-v1 weights for resnet34 and vgg16

This commit also adds the vgg16 model function. Before this, we only had
the modified vgg16 used for tumor detection.

* rm check that weights are TCGA-BRCA-v1

* fix inceptionv4-tils model sha256 + fix num classes for tils models

* add dockerfile with resnet34 weights for brca tumor

* set prad classes to benign, grade3, grade3+4

* update tcga luad v1 class names

lepidic, benign, acinar, micropapillary, mucinous, solid

* set TORCH_HOME to /var/lib/wsi_inference

* use TORCH_HOME from base image

* use a versioned base image

* set metadata as 87.5 microns in tcga-brca-v1

* fix til model patch sizes

Use 100px at 0.5 mpp. The patches are 50 micrometers.

* rm prad and luad models (and til vgg and resnet34)

These models are not ready for primetime.