FIX: conversion script to make SBU BMI output format (#37)
* add write_color_txt

Add function to write the color text file. This function takes a few
minutes per slide... This will likely double the runtime of the script,
but it seems to be necessary.

* save sbu bmi directory structure

* apply to dir of outputs + fix json heatmaps

* write meta json and txt files

* save run metadata to run_metadata.json

This saves a JSON file with lots of useful info about the run and the
envirnoment.

* use info from run_metadata.json

* write metadata json after inference

* use code cwd for git calls

* test for uncommitted changes

* save class-specific txt files

Also print run metadata and class names to console.

* fix prediction- headers + print info + fix txt save paths

* update sbu bmi conversion example

and move model conversion info to scripts/ dir.