ENH: update app dockerfiles + add mypy/flake8 to github ci + support weights from timm (#51)
* use image tag kaczmarj/wsinfer

* use workdir /opt/wsinfer and use default cmd --help

* add shields for python versions and pypi version

* add version key to yaml configs

* load weights from 'state_dict' key if it exists

When using timm for model training, weights are saved to the state_dict
key of the state dict file. This commit uses the parameters in the
state_dict key if it exists.

* move patching cli to wsinfer.cli namespace

* rm unused imports + split long lines

* ignore types for missing imports

* split long comment

* rm pretrained_settings dict

* add type annots to __getitem__

* move patching cli to wsinfer.cli namespace

* fix forward pass types for mypy

It seems that if I use the same variable name for the tensor or pil
image throughout the function definition, the Union type does not get
restricted to a Tensor type. But if I change the name of the variable,
mypy passes.

* run flake8 and mypy

* use Weights().load_model() to download weights

* add instructions to list available models

* add types- packages to dev extras

* unindent a line