Support MPS on Apple Silicon
See https://pytorch.org/docs/stable/notes/mps.html

Testing this with one image using a Mac Studio (M1 Max) the processing time decreased from ~21 minutes to ~2 minutes.

I'm not sure if the change in `utils.py` is necessary, nor if there should be changes in other places - it looks like the report in `infer.py` should be updated if this is accepted.

Also updated `.gitignore` for more Mac (and PyCharm) friendliness.