fix x,y coord creation + sort coords after query
Before this commit, we were missing the bottom and right edge of patches
because the stop of the coord range was set slightly too low. This
commit fixes that, so now all patches are represented.

This commit also sorts the coordinate indexes after shapely tree query,
because it seems that the order query returns is arbitrary. Sorting
that ensures that when we index the coordinates that are contained in
the polygon, those coordinates are in ascending order (with y changing
most rapidly).