[REF] use custom patching implementation and use Apache 2.0 license (#153)
* rm clam-based patch implementation

This implementation is licensed GPL3 and is incompatible with Apache 2.0
license. We would like to use Apache 2.0 so this commit removes the GPL3
licensed code.

* add histolab for patching impl

* replace histolab with custom impl

* add scikit-image >= 0.20.0

* add custom segment and patch impl

* replace cli with new patch impl

* use new patching impl

* use Apache 2.0 license

We have removed code from CLAM and are no longer bound by GPL. As of
this commit, wsinfer is licensed under Apache 2.0.

* when thresholding, use arr>=thresh

* use binary thresholding in lieu of otsu

* use new cli args in patching

* fix x,y coord creation + sort coords after query

Before this commit, we were missing the bottom and right edge of patches
because the stop of the coord range was set slightly too low. This
commit fixes that, so now all patches are represented.

This commit also sorts the coordinate indexes after shapely tree query,
because it seems that the order query returns is arbitrary. Sorting
that ensures that when we index the coordinates that are contained in
the polygon, those coordinates are in ascending order (with y changing
most rapidly).

* run isort

* mypy ignore skimage.morphology

* update expected number of patches in test-package

* add test patching test of 100px @ 0.5 mpp

* add binary_threshold argument

* remove patches that extend out of the border of the slide

Our patching algorithm ignores patches that go outside of the slide
boundary.