clone batch_coords tensor to prevent Too many open files error
When running wsinfer in a windows subsystem for linux environment, i got the error below. Appending a tensor from the dataloader to a list can cause this issue, and cloning the tensor solves it. This commit clones the tensor.

```
RuntimeError: Too many open files. Communication with the workers is no longer possible. Please increase the limit using `ulimit -n` in the shell or change the sharing strategy by calling `torch.multiprocessing.set_sharing_strategy('file_system')` at the beginning of your code
```