only use specified backend (or tifffile) to read mpp
Previously, even if tiffslide was the chosen backend, openslide
would still be used first to get mpp. This could cause panics
if a slide could not be read by openslide. This panic could
confuse users because they asked to use tiffslide. This commit
changes the mpp reading function to use the chosen backend or
tifffile if that backend fails.